third_party/blink/web_tests/platform/mac-mac10.11/external/wpt/html/rendering/widgets/button-layout/grid-expected.txt

This is a testharness.js-based test.
FAIL inline-grid assert_equals: clientWidth expected 28 but got 21
FAIL grid assert_equals: clientWidth expected 28 but got 21
Harness: the test ran to completion.
